Changed decode! syntax, uint to usize, box syntax, {:?}

+++ src/frame/stream/v2.rs
@@ -14,7 +14,7 @@ impl FrameStream for FrameV2 {
         let sizebytes = try!(reader.read_exact(3));
         let read_size = ((sizebytes[0] as u32) << 16) | ((sizebytes[1] as u32) << 8) | sizebytes[2] as u32;
 
-        let data = try!(reader.read_exact(read_size as uint));
+        let data = try!(reader.read_exact(read_size as usize));
         try!(frame.parse_data(data.as_slice()));
 
         Ok(Some((6 + read_size, frame)))
